backend/user/serializers/users.py
from rest_framework import serializers
from user.models.users import CustomUser
# from django.contrib.auth.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = '__all__'