backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2